requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.14.2
billiard==4.2.3
boto3==1.41.2
botocore==1.41.2
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==4.2.26
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
Events==0.5
greenlet==3.2.4
grpcio==1.76.0
idna==3.11
importlib_metadata==8.5.0
jmespath==1.0.1
kombu==5.5.4
Markdown==3.7
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
packaging==25.0
playwright==1.56.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psycopg2-binary==2.9.10
pyee==13.0.0
python-dateutil==2.9.0.post0
redis==7.1.0
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zipp==3.20.2
